feat(backup): Enhance backup script and add notifications
Adds a `restic check` to the beginning of the backup script to ensure
repository integrity before writing new data.

Implements backup notification using `ntfy`: a success notification is
sent upon completion, and a systemd `OnFailure` hook triggers a separate
service to send a failure notification.

`ntfy` is hosted on the server, so it's not as robust as it could be,
but I'll notice quickly if I stop getting backup success notifications.

Adds documentation with instructions for performing full or partial restores.