immich: do not override custom Exec of db container
Immich relies on the database having some custom options passed in
via the command line in the container's Entrypoint
setting `-c listen_addresses` in quadlet Exec will break immich[0]

There is no environment variable to set this, and while we could
set a custom Exec including the container's necessary arguments,
it would be fragile if they ever change.

This commit removes listen_addresses because it's not really required
on a postgres database secured with a password available only to
the immich application container. Even if some other container
on the network is compromised and is able to send requests to
Immich's DB, they will not get past the authentication check.

[0] See:
https://github.com/immich-app/immich/discussions/18674
https://github.com/immich-app/immich/issues/18432